[초이] add init.js, app.js, router.js
